Add Ansible install via rpm_path install on Cent7
Overrides the default of doing a .repo based install of ODL, which is
recommended for most use-cases. Uses the rpm_path install method and
accepts the default path provided by ansible-odl vars/main.yml.

A playbook is added with these options set and the previous playbook is
renamed for clarity.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>